Add periodIsLate function in 396.js
A new function called 'periodIsLate' has been added to the 396.js file. This function calculates the number of days between two given dates 'last' and 'today', and returns true if the difference is greater than a provided 'cycleLength'; otherwise false. This would be useful for tracking menstrual cycles or any repeated events.