Add logicalCalc function to handle logical calculations
This commit introduces a new function, logicalCalc, which takes an array and an operator (AND, OR, XOR) as arguments. The aim of this function is to reduce the array based on the given operator. This will add a utility to perform such logical operations that may be required in the application.