Add countSquares function in 403.js
A new file, 403.js, was created which included the function countSquares. The countSquares function is used to calculate the number of squares that can be made from cuts in a cube, as per the requirement from the related codewars challenge (link included in the comments). Additionally, it ensures that if no cuts are made, the result would be 1, as there is only one cuboid present.